Removed Iterator#remove
Removed the possibility to remove elements using Iterator. It is too
much of a task to figure out all the states in order to make in fully
functional in a concurrent environment. Also, I don't see any point in
having it... removing elements whereever in a queue isn't really an
attribute a queue should have.

Also fixed some other things...